requirements.txt
python-telegram-bot==20.7
pytesseract==0.3.10
Pillow==10.2.0
openpyxl==3.1.2
